refactor (frontend): unify wallet provider api
