+++ packages/frontend/src/components/contestedResources/contenders/ContendersList.js
@@ -1,11 +1,11 @@
 import { ContendersContent } from './Content'
 import { ContendersTemplate } from './Template'
 
-import { useWalletConnect } from '../../../hooks/useWallet'
+import { useWallet } from '../../../contexts'
 import { useVoteValidation } from './useVoteValidation'
 
 const ContendersList = ({ className, isFinished, ...props }) => {
-  const wallet = useWalletConnect()
+  const wallet = useWallet()
   const { isVoteVisible, prevVote, voteValidateState } = useVoteValidation({ wallet, isFinished })
   return (
         <ContendersTemplate isVoteVisible={isVoteVisible} className={className}>
